pyproject.toml
[project]
name = "radiofrequency"
version = "0.1.0"
description = "Generic RF signal reverse-engineering and replay toolkit"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "matplotlib>=3.10.8",
    "pytest>=8.0",
]

[project.scripts]
fanctl = "cli:cli"
